requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.32
pymysql>=1.1.1
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
PyJWT>=2.8.0
google-generativeai>=0.7.2
email-validator>=2.1.0
